Limits display field to 10 digits
This limits the user from entering more than 10 digits (not including the decimal character), and limits the display from showing more than 10 digits in the result. The displayed number is the one used in mathematical operations, so will be rounded for the next operation. This means that a trailing 0 will be displayed if an operation results in truncating a decimal value after a 0 character, but an operation on this result which does not change that decimal value or add digits to the left of the decimal will result in the next output leaving that 0 off as no longer significant. Numbers are handled as doubles such that these values should be handled correctly.